Added ToolbarTitle? test
